requirements.txt
tensorflow==2.14.0
numpy==1.24.3
opencv-python==4.8.0.74
pillow==10.0.0
scikit-learn==1.3.0
scikit-image==0.21.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.2
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
mlflow==2.8.0
prefect==2.13.0
prefect-dask==0.2.9
dvc==3.36.1
dvc-local==3.0.0
evidently==0.4.16
pytest==7.4.0
pytest-cov==4.1.0
httpx==0.24.1
requests==2.31.0
pyyaml==6.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
Jinja2==3.1.2
